I streamlined the subarray_threshold_scans and trigger_veto_window_tests_internal notebooks. All the firmware validation tests are in trigger_veto_window_tests_internal and the ethernet test notebook, and the threshold scan notebook focuses on observing the RFI environment with a subarray.  I created cr_data_inspection_functions.py to separate the handy quick plot tools from cr_functions.py which is for controlling the SNAP2s.
